// src/app/core/models/masters/general/location.ts
export interface Location {
    id: number;
    title: string;
    status: string;
    status_id :number;
    status_label: string;
}

export interface LocationPaginationResponse {
    current_page: number;
    total: number;
    per_page: number;
    last_page: number;
    from: number;
    data: Location[];
}